WIP: everything broken
I started splitting up the routes for simpler apis.

The idea is to have one route per pair.

Current state is that the original flow is broken because I rely on source_token which is BtcArkade and I use TokenId as input for the price estimate. Within price estimate I need to get the token contract. Given BtcArkade doesn't have a token contract, I error.
This is a simple fix, I hope afterwards the flow works again.

Afterwards

1. polygon - arkade flow seems to be plugged together, client side is completely missing
2. polygon - lightning flow needs a change in our rust sdk - client side is also missing